requirements.txt
asgiref==3.8.1
build==1.2.1
click==8.1.7
Django==5.0.4
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.0
pip-tools==7.4.1
pluggy==1.5.0
PyJWT==2.8.0
psycopg2-binary==2.9.3
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
setuptools==69.5.1
sqlparse==0.5.0
uvicorn==0.15.0
wheel==0.43.0